+++ checker-qual/src/main/java/org/checkerframework/checker/calledmethods/qual/RequiresCalledMethods.java
@@ -49,7 +49,7 @@ public @interface RequiresCalledMethods {
   @Documented
   @Retention(RetentionPolicy.RUNTIME)
   @Target({ElementType.METHOD, ElementType.CONSTRUCTOR})
-  @interface List {
+  public static @interface List {
     /**
      * Returns the repeatable annotations.
      *