+++ docs/developer/release/release_utils.py
@@ -24,7 +24,7 @@ from release_vars import execute
 # Parse Args Utils # TODO: Perhaps use argparse module
 
 
-def read_command_line_option(argv, argument):
+def has_command_line_option(argv, argument):
     """Returns True if the given command line arguments contain the specified
     argument, False otherwise."""
     for index in range(1, len(argv)):
@@ -130,7 +130,7 @@ def check_tools(tools):
     print(", ".join(tools))
     print(
         (
-            "Note: If you are NOT working on buffalo.cs.washington.edu then you "
+            "Note: If you are NOT working on the CSE file system then you "
             + "likely need to change the variables that are set in release.py\n"
             + 'Search for "Set environment variables".'
         )
@@ -260,16 +260,16 @@ def update_repo(path, bareflag):
     """Pull the latest changes to the given repo and update. The bareflag
     parameter indicates whether the updated repo must be a bare git repo."""
     if bareflag:
-        execute("git fetch origin master:master --ff-only", working_dir=path)
+        execute("git fetch origin master:master", working_dir=path)
     else:
-        execute("git pull", working_dir=path)
+        execute("git pull --ff-only", working_dir=path)
 
 
 def commit_tag_and_push(version, path, tag_prefix):
     """Commit the changes made for this release, add a tag for this release, and
     push these changes."""
     # Do nothing (instead of erring) if there is nothing to commit.
-    if execute('git diff-index --quiet HEAD', False, False, working_dir=path) == 0:
+    if execute("git diff-index --quiet HEAD", False, False, working_dir=path) == 0:
         execute('git commit -a -m "new release %s"' % (version), working_dir=path)
     execute("git tag %s%s" % (tag_prefix, version), working_dir=path)
     push_changes(path)
@@ -321,7 +321,7 @@ def is_repo_cleaned_and_updated(repo):
     and up-to-date with respect to the repository it was cloned from."""
     # The idiom "not execute(..., capture_output=True)" evaluates to True when the captured output is empty.
     if git_bare_repo_exists_at_path(repo):
-        execute("git fetch origin --ff-only", working_dir=repo)
+        execute("git fetch origin", working_dir=repo)
         is_updated = not execute(
             "git diff master..FETCH_HEAD", working_dir=repo, capture_output=True
         )
@@ -331,7 +331,7 @@ def is_repo_cleaned_and_updated(repo):
         is_clean = not execute(
             "git status --porcelain", working_dir=repo, capture_output=True
         )
-        execute("git fetch origin --ff-only", working_dir=repo)
+        execute("git fetch origin", working_dir=repo)
         is_updated = not execute(
             "git diff origin/master..master", working_dir=repo, capture_output=True
         )